# colbuilder: pool materializers if possible when wrapping row sources