Before 999533f, we would always put the
materializers needed to row-sourcify the inputs (when wrapping
a row-execution processor into the vectorized flow) into the
corresponding pool. However, that commit introduced a condition when
releasing the materializers back to the pool was unsafe, so we disabled
the optimization. This commit takes the optimization back partially to
explicitly control whether it is safe to pool those materializers.

Release note: None