@yuzefovich 2021-04-01
This change gives us the following minor improvement on a microbenchmark:
```
name                                                                                 old time/op    new time/op    delta
Planning/Cockroach/SELECT_*_FROM_abc-24                                                 385µs ± 4%     384µs ± 2%    ~     (p=1.000 n=8+7)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-24                          408µs ± 2%     405µs ± 2%    ~     (p=0.481 n=8+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_b_=_5-24                                     469µs ± 2%     452µs ± 4%  -3.73%  (p=0.000 n=9+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-24                          478µs ± 2%     461µs ± 3%  -3.46%  (p=0.001 n=9+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_c_=_5-24                                     455µs ± 3%     455µs ± 3%    ~     (p=0.959 n=8+8)
Planning/Cockroach/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-24              488µs ± 2%     472µs ± 4%  -3.33%  (p=0.006 n=8+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc-24                                        863µs ± 2%     847µs ± 3%    ~     (p=0.128 n=7+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-24                 716µs ± 2%     713µs ± 2%    ~     (p=0.489 n=9+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_b_=_5-24                            768µs ± 2%     777µs ± 3%    ~     (p=0.121 n=8+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-24                 785µs ± 2%     793µs ± 2%    ~     (p=0.165 n=7+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_c_=_5-24                            842µs ± 1%     773µs ± 2%  -8.22%  (p=0.000 n=6+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-24    1.11ms ± 3%    1.02ms ± 4%  -7.66%  (p=0.000 n=9+9)

name                                                                                 old alloc/op   new alloc/op   delta
Planning/Cockroach/SELECT_*_FROM_abc-24                                                25.4kB ± 2%    25.6kB ± 2%    ~     (p=0.295 n=8+7)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-24                         24.1kB ± 1%    24.1kB ± 1%    ~     (p=0.963 n=8+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_b_=_5-24                                    47.6kB ± 1%    45.0kB ± 0%  -5.51%  (p=0.000 n=8+8)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-24                         48.3kB ± 1%    46.0kB ± 1%  -4.83%  (p=0.000 n=8+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_c_=_5-24                                    47.1kB ± 0%    44.7kB ± 1%  -5.16%  (p=0.001 n=7+7)
Planning/Cockroach/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-24             53.0kB ± 0%    50.3kB ± 1%  -5.10%  (p=0.000 n=9+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc-24                                       77.8kB ± 2%    77.8kB ± 3%    ~     (p=0.955 n=7+8)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-24                41.2kB ± 4%    41.1kB ± 4%    ~     (p=0.863 n=9+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_b_=_5-24                           64.2kB ± 2%    62.0kB ± 4%  -3.29%  (p=0.026 n=7+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-24                65.2kB ± 2%    62.9kB ± 4%  -3.61%  (p=0.011 n=7+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_c_=_5-24                           64.7kB ± 3%    61.8kB ± 1%  -4.47%  (p=0.001 n=7+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-24     119kB ± 3%     115kB ± 3%  -2.67%  (p=0.000 n=9+9)

name                                                                                 old allocs/op  new allocs/op  delta
Planning/Cockroach/SELECT_*_FROM_abc-24                                                   241 ± 0%       243 ± 0%  +0.70%  (p=0.003 n=7+6)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-24                            235 ± 0%       235 ± 0%    ~     (p=0.316 n=8+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_b_=_5-24                                       366 ± 0%       361 ± 0%  -1.49%  (p=0.000 n=9+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-24                            374 ± 0%       370 ± 0%  -1.22%  (p=0.000 n=9+9)
Planning/Cockroach/SELECT_*_FROM_abc_WHERE_c_=_5-24                                       361 ± 0%       356 ± 1%  -1.20%  (p=0.000 n=7+8)
Planning/Cockroach/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-24                397 ± 0%       391 ± 1%  -1.43%  (p=0.000 n=9+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc-24                                          767 ± 0%       764 ± 1%  -0.40%  (p=0.046 n=7+8)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-24                   410 ± 2%       409 ± 1%    ~     (p=0.393 n=9+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_b_=_5-24                              541 ± 1%       536 ± 1%  -0.99%  (p=0.014 n=8+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-24                   550 ± 1%       542 ± 1%  -1.35%  (p=0.003 n=6+7)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_WHERE_c_=_5-24                              539 ± 1%       530 ± 1%  -1.63%  (p=0.000 n=7+9)
Planning/MultinodeCockroach/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-24       991 ± 1%       983 ± 1%  -0.87%  (p=0.002 n=9+9)
```